+++ lib/toggle_switch.dart
@@ -64,6 +64,9 @@ class ToggleSwitch extends StatefulWidget {
   /// Icon size
   final double iconSize;
 
+  /// Divider margin
+  final double? dividerMargin;
+
   /// Border width
   final double? borderWidth;
 
@@ -97,42 +100,43 @@ class ToggleSwitch extends StatefulWidget {
   /// Use toggle switch vertically
   final bool isVertical;
 
-  // Set a border only to the active toggle component
-  Border? activeBorder;
+  /// Set a border only to the active toggle component
+  List<Border?>? activeBorders;
 
   ToggleSwitch(
       {Key? key,
-        this.totalSwitches,
-        this.labels,
-        this.borderColor,
-        this.borderWidth,
-        this.dividerColor = Colors.white30,
-        this.activeBgColor,
-        this.activeFgColor,
-        this.inactiveBgColor,
-        this.inactiveFgColor,
-        this.onToggle,
-        this.cornerRadius = 8.0,
-        this.initialLabelIndex = 0,
-        this.minWidth = 72.0,
-        this.minHeight = 40.0,
-        this.changeOnTap = true,
-        this.icons,
-        this.activeBgColors,
-        this.customTextStyles,
-        this.customIcons,
-        this.customWidths,
-        this.customHeights,
-        this.animate = false,
-        this.animationDuration = 800,
-        this.curve = Curves.easeIn,
-        this.radiusStyle = false,
-        this.textDirectionRTL = false,
-        this.fontSize = 14.0,
-        this.iconSize = 17.0,
-        this.doubleTapDisable = false,
-        this.isVertical = false,
-        this.activeBorder})
+      this.totalSwitches,
+      this.labels,
+      this.borderColor,
+      this.borderWidth,
+      this.dividerColor = Colors.white30,
+      this.activeBgColor,
+      this.activeFgColor,
+      this.inactiveBgColor,
+      this.inactiveFgColor,
+      this.onToggle,
+      this.cornerRadius = 8.0,
+      this.initialLabelIndex = 0,
+      this.minWidth = 72.0,
+      this.minHeight = 40.0,
+      this.changeOnTap = true,
+      this.icons,
+      this.activeBgColors,
+      this.customTextStyles,
+      this.customIcons,
+      this.customWidths,
+      this.customHeights,
+      this.animate = false,
+      this.animationDuration = 800,
+      this.curve = Curves.easeIn,
+      this.radiusStyle = false,
+      this.textDirectionRTL = false,
+      this.fontSize = 14.0,
+      this.iconSize = 17.0,
+      this.dividerMargin = 8.0,
+      this.doubleTapDisable = false,
+      this.isVertical = false,
+      this.activeBorders})
       : super(key: key);
 
   @override
@@ -166,20 +170,27 @@ class _ToggleSwitchState extends State<ToggleSwitch>
   @override
   Widget build(BuildContext context) {
     /// Searches for largest totalSwitches value
-    final int totalSwitches = widget.totalSwitches ?? [widget.labels?.length ?? 0, widget.icons?.length ?? 0, widget.customIcons?.length ?? 0].reduce(max);
+    final int totalSwitches = widget.totalSwitches ??
+        [
+          widget.labels?.length ?? 0,
+          widget.icons?.length ?? 0,
+          widget.customIcons?.length ?? 0
+        ].reduce(max);
     super.build(context);
 
     /// Assigns active background color to default primary theme color if it's null/not provided.
     activeBgColor = widget.activeBgColor ?? [Theme.of(context).primaryColor];
 
     /// Assigns active foreground color to default accent text theme color if it's null/not provided.
-    activeFgColor = widget.activeFgColor ?? Theme.of(context).primaryTextTheme.bodyText1!.color;
+    activeFgColor = widget.activeFgColor ??
+        Theme.of(context).primaryTextTheme.bodyText1!.color;
 
     /// Assigns inactive background color to default disabled theme color if it's null/not provided.
     inactiveBgColor = widget.inactiveBgColor ?? Theme.of(context).disabledColor;
 
     /// Assigns inactive foreground color to default text theme color if it's null/not provided.
-    inactiveFgColor = widget.inactiveFgColor ?? Theme.of(context).textTheme.bodyText1!.color;
+    inactiveFgColor =
+        widget.inactiveFgColor ?? Theme.of(context).textTheme.bodyText1!.color;
 
     /// Assigns transparent border color if it's null/not provided.
     borderColor = widget.borderColor ?? [Colors.transparent];
@@ -188,7 +199,8 @@ class _ToggleSwitchState extends State<ToggleSwitch>
     ///
     /// Assigns default border width of 0.0 if border color is null.
     /// Assigns default border width of 3.0 if border width is null but border color is not.
-    borderWidth = widget.borderWidth ?? (widget.borderColor == null ? 0.0 : 3.0);
+    borderWidth =
+        widget.borderWidth ?? (widget.borderColor == null ? 0.0 : 3.0);
 
     return ClipRRect(
       borderRadius: BorderRadius.circular(widget.cornerRadius),
@@ -242,10 +254,17 @@ class _ToggleSwitchState extends State<ToggleSwitch>
 
                 /// Returns item divider
                 return Container(
-                  width: !widget.isVertical ? 1 : widget.minWidth,
+                  width: !widget.isVertical ? 1 : double.infinity,
                   height: widget.isVertical ? 1 : double.infinity,
-                  color: activeDivider ? Colors.transparent : widget.dividerColor,
-                  margin: widget.isVertical ? EdgeInsets.zero : EdgeInsets.symmetric(vertical: activeDivider ? 0 : 8),
+                  color:
+                      activeDivider ? Colors.transparent : widget.dividerColor,
+                  margin: widget.isVertical
+                      ? EdgeInsets.symmetric(
+                          horizontal:
+                              activeDivider ? 0.0 : widget.dividerMargin!)
+                      : EdgeInsets.symmetric(
+                          vertical:
+                              activeDivider ? 0.0 : widget.dividerMargin!),
                 );
               } else {
                 /// Matches corner radius of active switch to that of border
@@ -254,45 +273,63 @@ class _ToggleSwitchState extends State<ToggleSwitch>
                   /// Checks if text direction is set right-to-left and
                   /// assigns corner radius accordingly.
                   cornerRadius = widget.textDirectionRTL
-                      ? BorderRadius.horizontal(right: Radius.circular(widget.cornerRadius),)
-                      : BorderRadius.horizontal(left: Radius.circular(widget.cornerRadius),);
+                      ? BorderRadius.horizontal(
+                          right: Radius.circular(widget.cornerRadius),
+                        )
+                      : BorderRadius.horizontal(
+                          left: Radius.circular(widget.cornerRadius),
+                        );
                 }
                 if (index == 0 && widget.isVertical) {
-                  cornerRadius = BorderRadius.vertical(top: Radius.circular(widget.cornerRadius),);
+                  cornerRadius = BorderRadius.vertical(
+                    top: Radius.circular(widget.cornerRadius),
+                  );
                 }
                 if (index ~/ 2 == totalSwitches - 1 && !widget.isVertical) {
                   /// Checks if text direction is set right-to-left and
                   /// assigns corner radius accordingly.
                   cornerRadius = widget.textDirectionRTL
-                      ? BorderRadius.horizontal(left: Radius.circular(widget.cornerRadius),)
-                      : BorderRadius.horizontal(right: Radius.circular(widget.cornerRadius),);
+                      ? BorderRadius.horizontal(
+                          left: Radius.circular(widget.cornerRadius),
+                        )
+                      : BorderRadius.horizontal(
+                          right: Radius.circular(widget.cornerRadius),
+                        );
                 }
                 if (index ~/ 2 == totalSwitches - 1 && widget.isVertical) {
-                  cornerRadius = BorderRadius.vertical(bottom: Radius.circular(widget.cornerRadius),);
+                  cornerRadius = BorderRadius.vertical(
+                    bottom: Radius.circular(widget.cornerRadius),
+                  );
                 }
 
                 /// Assigns empty widget if icon is null
                 /// Calculates icon's size to prevent overflow
-                var icon =
-                widget.icons != null && widget.icons![index ~/ 2] != null
+                var icon = widget.icons != null &&
+                        widget.icons![index ~/ 2] != null
                     ? Icon(
-                  widget.icons![index ~/ 2],
-                  color: fgColor,
-                  size: widget.isVertical ?
-                  widget.iconSize >
-                      (_calculateHeight(index ~/ 2, totalSwitches) / 3)
-                      ? (_calculateHeight(index ~/ 2, totalSwitches)) / 3
-                      : widget.iconSize
-                      :
-                  widget.iconSize >
-                      (_calculateWidth(index ~/ 2, totalSwitches) / 3)
-                      ? (_calculateWidth(index ~/ 2, totalSwitches)) / 3
-                      : widget.iconSize,
-                )
+                        widget.icons![index ~/ 2],
+                        color: fgColor,
+                        size: widget.isVertical
+                            ? widget.iconSize >
+                                    (_calculateHeight(
+                                            index ~/ 2, totalSwitches) /
+                                        3)
+                                ? (_calculateHeight(
+                                        index ~/ 2, totalSwitches)) /
+                                    3
+                                : widget.iconSize
+                            : widget.iconSize >
+                                    (_calculateWidth(
+                                            index ~/ 2, totalSwitches) /
+                                        3)
+                                ? (_calculateWidth(index ~/ 2, totalSwitches)) /
+                                    3
+                                : widget.iconSize,
+                      )
                     : Container();
 
-                /// Assigns custom icon if available
-                /// Overwrites icons passed via icons:
+                /// Assigns custom icon if available.
+                /// Overrides icons passed via icons:
                 if (widget.customIcons != null &&
                     widget.customIcons![index ~/ 2] != null) {
                   icon = widget.customIcons![index ~/ 2]!;
@@ -300,13 +337,35 @@ class _ToggleSwitchState extends State<ToggleSwitch>
 
                 /// Assigns custom text styles if available.
                 /// Assigns default text style if custom text style is not available.
-                var textStyle = widget.customTextStyles != null &&
-                    widget.customTextStyles![index ~/ 2] != null
-                    ? widget.customTextStyles![index ~/ 2]
-                    : TextStyle(
+                /// Overrides fontSize, activeFgColor, inactiveFgColor.
+                /// If only one TextStyle is passed then we assume that we wanna
+                /// apply that TextStyle to all the switches.
+                TextStyle defaultTextStyle = TextStyle(
                   color: fgColor,
                   fontSize: widget.fontSize,
                 );
+                var textStyle = defaultTextStyle;
+                if (widget.customTextStyles != null) {
+                  textStyle = widget.customTextStyles!.length == 1
+                      ? widget.customTextStyles![0]!
+                      : (widget.customTextStyles!.length > index ~/ 2 &&
+                              widget.customTextStyles![index ~/ 2] != null
+                          ? widget.customTextStyles![index ~/ 2]!
+                          : defaultTextStyle);
+                }
+
+                /// Assigns active border if available.
+                /// If only one active border is passed then we assume that we wanna
+                /// apply that active border to all the switches.
+                var activeBorder;
+                if (widget.activeBorders != null) {
+                  activeBorder = widget.activeBorders!.length == 1
+                      ? widget.activeBorders![0]
+                      : (widget.activeBorders!.length > index ~/ 2 &&
+                              widget.activeBorders![index ~/ 2] != null
+                          ? widget.activeBorders![index ~/ 2]!
+                          : null);
+                }
 
                 /// Returns switch item
                 return GestureDetector(
@@ -314,14 +373,19 @@ class _ToggleSwitchState extends State<ToggleSwitch>
                   child: AnimatedContainer(
                     padding: const EdgeInsets.only(left: 10.0, right: 10.0),
                     constraints: BoxConstraints(
-                      maxWidth: widget.isVertical ? BoxConstraints().maxWidth : _calculateWidth(index ~/ 2, totalSwitches),
-                      maxHeight: widget.isVertical ? _calculateHeight(index ~/ 2, totalSwitches) : BoxConstraints().maxHeight,
+                      maxWidth: widget.isVertical
+                          ? BoxConstraints().maxWidth
+                          : _calculateWidth(index ~/ 2, totalSwitches),
+                      maxHeight: widget.isVertical
+                          ? _calculateHeight(index ~/ 2, totalSwitches)
+                          : BoxConstraints().maxHeight,
                     ),
                     alignment: Alignment.center,
                     decoration: BoxDecoration(
-                      border: active ? widget.activeBorder : null,
+                      border: active ? activeBorder : null,
                       borderRadius: widget.radiusStyle
-                          ? BorderRadius.all(Radius.circular(widget.cornerRadius))
+                          ? BorderRadius.all(
+                              Radius.circular(widget.cornerRadius))
                           : cornerRadius,
                       gradient: LinearGradient(
                         colors: bgColor!.length == 1
@@ -331,7 +395,9 @@ class _ToggleSwitchState extends State<ToggleSwitch>
                         end: Alignment.bottomRight,
                       ),
                     ),
-                    duration: Duration(milliseconds: widget.animate ? widget.animationDuration : 0),
+                    duration: Duration(
+                        milliseconds:
+                            widget.animate ? widget.animationDuration : 0),
                     curve: widget.curve,
                     child: Row(
                       mainAxisAlignment: MainAxisAlignment.center,
@@ -381,6 +447,7 @@ class _ToggleSwitchState extends State<ToggleSwitch>
   }
 
   /// Calculates width to prevent overflow by taking screen width into account.
+  /// Ignores customWidths if toggle switch is vertical
   double _calculateWidth(int index, int totalSwitches) {
     /// Extra width to prevent overflow and add padding
     double extraWidth = 0.10 * totalSwitches;
@@ -392,17 +459,14 @@ class _ToggleSwitchState extends State<ToggleSwitch>
     ///
     /// Returns passed minWidth per label if total requested width plus extra width is less than max screen width.
     /// Returns calculated width to fit within the max screen width if total requested width plus extra width is more than max screen width.
-    return (totalSwitches + extraWidth) *
-        (widget.customWidths != null
-            ? widget.customWidths![index]
-            : widget.minWidth) <
-        screenWidth
-        ? (widget.customWidths != null
+    return widget.customWidths != null
         ? widget.customWidths![index]
-        : widget.minWidth)
-        : screenWidth / (totalSwitches + extraWidth);
+        : ((totalSwitches + extraWidth) * widget.minWidth < screenWidth
+            ? widget.minWidth
+            : screenWidth / (totalSwitches + extraWidth));
   }
 
+  /// Ignores customHeights if toggle switch is horizontal
   double _calculateHeight(int index, int totalSwitches) {
     /// Extra height to prevent overflow and add padding
     double extraHeight = 0.10 * totalSwitches;
@@ -414,15 +478,11 @@ class _ToggleSwitchState extends State<ToggleSwitch>
     ///
     /// Returns passed minHeight per label if total requested width plus extra height is less than max screen height.
     /// Returns calculated width to fit within the max screen width if total requested width plus extra height is more than max screen height.
-    return (totalSwitches + extraHeight) *
-        (widget.customHeights != null
-            ? widget.customHeights![index]
-            : widget.minHeight) <
-        screenHeight
-        ? (widget.customHeights != null
+    return widget.customHeights != null
         ? widget.customHeights![index]
-        : widget.minHeight)
-        : screenHeight / (totalSwitches + extraHeight);
+        : ((totalSwitches + extraHeight) * widget.minHeight < screenHeight
+            ? widget.minHeight
+            : screenHeight / (totalSwitches + extraHeight));
   }
 }
 
@@ -435,13 +495,31 @@ class RowToColumn extends StatelessWidget {
   final MainAxisSize? mainAxisSize;
   final CrossAxisAlignment? crossAxisAlignment;
 
-  const RowToColumn({Key? key, required this.children, required this.isVertical, this.isColumnToRow,
-    this.mainAxisAlignment, this.mainAxisSize, this.crossAxisAlignment}) : super(key: key);
+  const RowToColumn(
+      {Key? key,
+      required this.children,
+      required this.isVertical,
+      this.isColumnToRow,
+      this.mainAxisAlignment,
+      this.mainAxisSize,
+      this.crossAxisAlignment})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return ((isColumnToRow ?? false) ? isVertical : !isVertical)?
-    Row(children: children,     mainAxisAlignment: mainAxisAlignment ?? Row().mainAxisAlignment, mainAxisSize: mainAxisSize ?? Row().mainAxisSize, crossAxisAlignment: crossAxisAlignment ?? Row().crossAxisAlignment,) :
-    Column(children: children,  mainAxisAlignment: mainAxisAlignment ?? Column().mainAxisAlignment, mainAxisSize: mainAxisSize ?? Column().mainAxisSize, crossAxisAlignment: crossAxisAlignment ?? Column().crossAxisAlignment,);
+    return ((isColumnToRow ?? false) ? isVertical : !isVertical)
+        ? Row(
+            children: children,
+            mainAxisAlignment: mainAxisAlignment ?? Row().mainAxisAlignment,
+            mainAxisSize: mainAxisSize ?? Row().mainAxisSize,
+            crossAxisAlignment: crossAxisAlignment ?? Row().crossAxisAlignment,
+          )
+        : Column(
+            children: children,
+            mainAxisAlignment: mainAxisAlignment ?? Column().mainAxisAlignment,
+            mainAxisSize: mainAxisSize ?? Column().mainAxisSize,
+            crossAxisAlignment:
+                crossAxisAlignment ?? Column().crossAxisAlignment,
+          );
   }
 }